完整评论功能 最后历时202天 完结，撒花!!!

+++ blog-backend/src/main/java/org/example/mapper/AccountMapper.java
@@ -15,7 +15,7 @@ public interface AccountMapper extends BaseMapper<Account> {
 
     int insertAccount(Account account);
 
-    int deleteAccount(Integer uid);
+    int deleteAccount(Long uid);
 
     List<Account> getAccountCount();
